Mostly untested checkpoint commit of modules changes.
Heavy changes, mostly not fully tested,
to the entire system of modules, specifically
deleting, and zip uploads.

Started work on still mostly undocumented
API for custom resource types.

The plan is most likely for the custom resource
API to remain internal for a little while so
it can be ironed out.