Fix various bugs including more bugs in saving and loading, misc
Modules with embedded percent-encoding in the names were not being
correctly decoded in some cases.

Things were being listed as unsaved, and there were errors relating to
the code to remove certain things from the unsaved list.

In addition, external modules with percent encoded names would cause
issues on saving.

Also, the external_module_locations dict was being modified outside of
moduleslock. That's a remnant from when the registry was being used for
that instead I think.